Add --namingConvention option to generator
Currently only one option is supported: js.

--namingConvention js will generate type names in PascalCase, other names in
camelCase. This can be especially useful in case of Hasura generated schemas
where all names default to snake_case.

The js naming convention also generates different root collection names: these
will be camelCased and properly pluralized. Without --namingConvention js the
root  names are all lowercase and and a plural "s" is appended.
The +"s" however is not always valid English plural in case the root type's
name is  eg. "EmptyBox" as it will become "emptyboxs".
With  --namingConvention js it will be "emptyBoxes".

If --namingConvention is not set or set to anything else than "js" then the
default names provided by the graphql schema will be used.

MSTGQLStore.ts and mergeHelper.ts have also been updated so that root collection
name generation uses the same transformation as at generation time. For this
to work configureStoreMixin() receives the optional namingConvention value as
the third parameter and store.getCollectionName() is used at runtime to convert
__typename to a root collection name.